refactor: update entity test folder structor
